refactor: move context menu to plugins
