+++ src/features/NoteEditor/RichEditor/plugins/ContextMenu/components/GenericContextMenu.tsx
@@ -3,8 +3,9 @@ import React, { FC } from 'react';
 import { $isLinkNode } from '@lexical/link';
 import { $findMatchingParent } from '@lexical/utils';
 
-import { $isImageNode } from '../plugins/Image/ImageNode';
-import { ContextMenuRendererProps } from './ContextMenu';
+import { $isImageNode } from '../../Image/ImageNode';
+
+import { ContextMenuRendererProps } from '../ContextMenuPlugin';
 import { ObjectPropertiesEditor } from './ObjectPropertiesEditor';
 
 export const GenericContextMenu: FC<ContextMenuRendererProps> = ({
